requirements.txt
requests
fake-useragent
python-dotenv
Pillow